- moved the tokenBalance entry on the accountExtra state - avoid fetiching unused data on the app details page - moved the account staking logic outside the account layout
