+++ src/hooks/fetch/index.ts
@@ -1,3 +1,4 @@
+export * from './useFetchAccountStakingDetails';
 export * from './useFetchEconomics';
 export * from './useFetchGlobalStake';
 export * from './useFetchMarkers';
